Рефакторит компоненты в зависимости от статуса авторизации
