+++ project/src/components/app/app.tsx
@@ -1,7 +1,7 @@
 import {Route, BrowserRouter, Routes} from 'react-router-dom';
 import {HelmetProvider} from 'react-helmet-async';
 import NotFoundScreen from '../../pages/not-found-screen//not-found-screen';
-import OfferScreenNotLogged from '../../pages/offer-screen-not-logged/offer-screen-not-logged';
+import OfferScreenNotLogged from '../../pages/offer-screen/offer-screen';
 import LoginScreen from '../../pages/login-screen/login-screen';
 import ScrollToTop from '../scroll-to-top/scroll-to-top';
 
@@ -9,7 +9,7 @@ import {AppRoute} from '../../const/app-route';
 
 import type { City} from '../../@types/offer-types';
 import { Review } from '../../@types/review-types';
-import MainScreenNotLogged from '../../pages/main-screen-not-logged/main-screen-not-logged';
+import MainScreenNotLogged from '../../pages/main-screen/main-screen';
 
 type AppProps = {
   cities: City[];
