+++ project/src/components/header/sign-out.tsx
@@ -1,13 +1,17 @@
 import { Link } from 'react-router-dom';
 import {AppRoute} from '../../const/app-route';
+import UserName from './user-name';
 
 function SignOut(): JSX.Element {
   return (
-    <li className="header__nav-item">
-      <Link className="header__nav-link" title="To the login page" to={AppRoute.Login}>
-        <span className="header__signout">Sign out</span>
-      </Link>
-    </li>
+    <>
+      <UserName />
+      <li className="header__nav-item">
+        <Link className="header__nav-link" title="To the login page" to={AppRoute.Login}>
+          <span className="header__signout">Sign out</span>
+        </Link>
+      </li>
+    </>
   );
 }
