+++ project/src/pages/main-screen/main-screen.tsx
@@ -1,18 +1,18 @@
 import MainContent from '../../components/main-content/main-content';
-import HeaderSignOut from '../../components/header/header-sign-out';
+import Header from '../../components/header/header';
 import {Helmet} from 'react-helmet-async';
 import type { MainScreenProps } from '../../components/main-content/main-content';
 
-function MainScreenLogged({cities}: MainScreenProps): JSX.Element {
+function MainScreenNotLogged({cities}: MainScreenProps): JSX.Element {
   return (
     <div className="page page--gray page--main">
       <Helmet>
-        <title>Шесть городов.Добро пожаловать!</title>
+        <title>Шесть городов.Требуется регистрация!</title>
       </Helmet>
-      <HeaderSignOut />
+      <Header />
       <MainContent cities={cities}/>
     </div>
   );
 }
 
-export default MainScreenLogged;
+export default MainScreenNotLogged;
